% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BendabilityProfile.R
\name{BendabilityProfile}
\alias{BendabilityProfile}
\title{Bendability profile of DNA sequence}
\usage{
BendabilityProfile(sequences, scale, k, plot.it = TRUE, smt = 0)
}
\arguments{
\item{sequences}{A DNAStringSet object or a character vector.}

\item{scale}{One from "con", "conrigid", "dnase", "dnaserigid", "nuc", "nucrigid".}

\item{k}{Number of consecutive trinucleotides for which to calculate the average
bendability coefficient.}

\item{plot.it}{Whether to plot the bendability profile(s) or not. Defaults to TRUE.}

\item{smt}{Degree of line smoothing. Between 0 and 1, defaults to 0 (no smoothing).}
}
\value{
A data.table, or a plot and a data.table. Each input sequence is represented
by one row in data.table, which contains average bendability coefficients for (k+2)-mers.
}
\description{
Splits sequence(s) into (k+2)-mers with two-nucleotide overlap and calculates average
bendability per (k+2)-mer according to the chosen scale. Plots the result.
}
\details{
Parameter \strong{k} is defined as number of consecutive trinucleotides rather
than number of nucleotides (or window size) because a bendability coefficient is
defined for a trinucleotide. Size of window is actually k+2 nucleotides.
}
\examples{
BendabilityProfile("TGATTCCTAAAGTCA", "con", 1, plot.it=TRUE)

s <- c("TGATTCCTAAAGTCA", "CCTGAAATGCTAGCGT", "AAACTAGCCTCGATG")
BendabilityProfile(s, "con", 1, plot.it=TRUE, smt=0.3)
BendabilityProfile(s, "con", 3, plot.it=FALSE)
}
