% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bendDNA.R
\docType{package}
\name{bendDNA}
\alias{bendDNA}
\alias{bendDNA-package}
\title{bendDNA: A package for DNA bendability calculations}
\description{
Package provides functions for:
\itemize{
 \item{calculating and plotting bendability profile of DNA sequence}
 \item{calculating average bendabilities of all possible k-mers}
 \item{finding all sequences with similar bendability profiles}
 \item{clustering sequences with similar bendabilities, and calculating their consensus}
}
}
\author{
Dunja Glavas, Michiel Coghe
}
