% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAndConsensus.R
\name{ClusterAndConsensus}
\alias{ClusterAndConsensus}
\title{Clustering and consensus of similar sequences}
\usage{
ClusterAndConsensus(out, cutoff, sample.seq = NULL)
}
\arguments{
\item{out}{A single data.table output of function MatchBendability
(output.list=FALSE), or any data.table containing sequences in a column named
"sequence".}

\item{cutoff}{Numeric, between 0 and 1. Distance required for separation of clusters.}

\item{sample.seq}{If it is NULL (default), all sequences from input will be used in
clustering and consensus calculation. If a single integer x is provided, x
input sequences will be chosen at random. Can also be a vector of indexes of
desired sequences (their row positions in the input data.table).}
}
\value{
Variable number of consensus sequences written in extended genetic alphabet
   (IUPAC_CODE_MAP).
}
\description{
Intended to present output of function MatchBendability in a more human-friendly
format. The output can contain very large numbers of highly similar sequences,
so clustering and calculating a consensus can help provide a clearer picture.
DISCLAIMER: due to the nature of bendability problem and limitations of IUPAC
extended genetic alphabet, output of this function is not intended for further machine
use. Not all sequences which can be inferred from consensus output will have satisfied
the initial bendability requirements (for example AA/TT will be marked as WW, same as
AT/TA).
}
\details{
Similar sequences are clustered together using inexact method
(sequence-only method, doesn't use a distance matrix). Number of clusters depends
on sequence similarity and cutoff value (lower values will produce more clusters).
}
\examples{
dt <- MatchBendability("TGATTCCTAAAGTCA", "con", k=3, tolerance=0.1)
ClusterAndConsensus(dt, cutoff=0.01)
ClusterAndConsensus(dt, cutoff=0.5)
}
