% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBendability.R
\name{PlotBendability}
\alias{PlotBendability}
\title{Plot bendability of sequence(s)}
\usage{
PlotBendability(sequences, scale, k, sample.seq = NULL, smt = 0)
}
\arguments{
\item{sequences}{A DNAStringSet object or a character vector of same-length DNA
sequences.}

\item{scale}{One from "con", "conrigid", "dnase", "dnaserigid", "nuc", "nucrigid".}

\item{k}{Number of consecutive trinucleotides for which to calculate and plot the
average bendability coefficient.}

\item{sample.seq}{If it is NULL (default), all sequences from input will be plotted.
If a single integer x is provided, x input sequences will be chosen at random.
Can also be a vector of indexes of desired sequences (their row positions in the
input data.table).}

\item{smt}{Whether to smooth the lines or not. Between 0 and 1, defaults to 0
(no smoothing).}
}
\value{
A plot.
}
\description{
Graphical representation of bendability profiles of one or several DNA sequences.
}
\examples{
PlotBendability("TGATTCCTAAAGTCA", "dnase", k=1)
PlotBendability(c("TGATTCCTAAAGTCA", "CCTGAAATGCTAGCGT", "AAACTAGCCTCGATG"), "dnase", k=3, smt=0.5)

dt <- MatchBendability("TGATTCCTAAAGTCA", "con", k=1, tolerance=1)
PlotBendability(dt$sequence, "dnase", k=3, sample.seq=10, smt=0.2)
}
