% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scales}
\alias{scales}
\title{Scales for trinucleotide bendability coefficients.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 64 rows and 7 columns.}
\usage{
scales
}
\description{
Available scales are "con", "conrigid", "dnase", "dnaserigid", "nuc", "nucrigid".
}
\keyword{datasets}
