% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BendabilityProfile.R
\name{BendabilityProfile}
\alias{BendabilityProfile}
\title{Bendability profile of DNA sequence}
\usage{
BendabilityProfile(query, scale, k)
}
\arguments{
\item{query}{DNAString or a character string.}

\item{scale}{One from "con", "conrigid", "dnase", "dnaserigid", "nuc", "nucrigid".}

\item{k}{Number of consecutive trinucleotides for which to calculate the average
bendability coefficient.}
}
\value{
A three-column data.table:
\itemize{
 \item{kmer} : {sequence of length k+2}
 \item{position} : {relative position of k-mer in query sequence}
 \item{avg.bend} : {average bendability of k-mer}
}
}
\description{
Splits sequence into k-mers with two-nucleotide overlap and calculates average
bendability per k-mer according to the chosen scale.
}
\details{
Parameter \strong{k} is defined as number of consecutive trinucleotides rather
than number of nucleotides (window size) because a bendability coefficient is
defined for a trinucleotide. Size of window is actually k+2 nucleotides.
}
\examples{
BendabilityProfile("TGATTCCTAAAGTCA", "con", 1)
BendabilityProfile("TGATTCCTAAAGTCA", "con", 3)
BendabilityProfile("TGATTCCTAAAGTCA", "nuc", 3)
}
