% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAndConsensus.R
\name{ClusterAndConsensus}
\alias{ClusterAndConsensus}
\title{Clustering and consensus of similar sequences}
\usage{
ClusterAndConsensus(sequences, cutoff, sample.seq = NULL)
}
\arguments{
\item{sequences}{A DNAStringSet object or a character vector of same-length DNA
sequences.}

\item{cutoff}{Numeric, between 0 and 1. Distance required for separation of clusters.}

\item{sample.seq}{If it is NULL (default), all sequences from input will be used in
clustering and consensus calculation. If a single integer x is provided, x
input sequences will be chosen at random. Can also be a vector of desired sequences'
indexes in the input.}
}
\value{
A data.table containing variable number of consensus sequences written in
   IUPAC's extended genetic alphabet.
}
\description{
Intended to present output of function MatchBendability in a more read-friendly
format. The output can contain very large numbers of highly similar sequences,
so clustering and calculating a consensus can help provide a clearer picture.
DISCLAIMER: due to the nature of bendability problem and limitations of extended
genetic alphabet, output of this function is not intended for further machine use.
Not all sequences which can be inferred from consensus output will have satisfied
the initial bendability requirements (for example AA/TT will be marked as WW, same as
AT/TA).
}
\details{
Similar sequences are clustered together using inexact method
(sequence-only method, doesn't use a distance matrix). Number of clusters depends
on sequence similarity and cutoff value (lower values will produce more clusters).

For the extended genetic alphabet, see \code{\link[Biostrings]{IUPAC_CODE_MAP}}.
}
\examples{
dt <- MatchBendability("TGATTCCTAAAGTCA", "con", k=3, tolerance=0.1)
ClusterAndConsensus(dt$sequence, cutoff=0.25)
ClusterAndConsensus(dt$sequence, cutoff=0.85)
}
