% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookupTable.R
\name{LookupTable}
\alias{LookupTable}
\title{Average bendability coefficients of all possible k-mers}
\usage{
LookupTable(scale, k, sequence.out = F)
}
\arguments{
\item{scale}{One from "con", "conrigid", "dnase", "dnaserigid", "nuc", "nucrigid".}

\item{k}{Number of consecutive trinucleotides for which to calculate the average
bendability coefficient.}

\item{sequence.out}{Whether to output (k+2)-mer sequences (TRUE) or indexes of
alphabetically sorted permutations (FALSE). Defaults to FALSE.}
}
\value{
A four-column data.table:
\itemize{
 \item{sequence} : {sequence of length k+2}
 \item{index} : {alternatively, index of permutation}
 \item{Lbend} : {average bendability}
 \item{pref} : {first two nucleotides of sequence/permutation}
 \item{suff} : {last two nucleotides of sequence/permutation}
}
}
\description{
Produces all possible (k+2)-mers for a given k and calculates their average bendability
coefficients. Upper limit for k is determined by the size of available RAM.
}
\details{
Parameter \strong{k} is defined as number of consecutive trinucleotides rather
than number of nucleotides (or window size) because a bendability coefficient is
defined for a trinucleotide. Size of such window is actually k+2 nucleotides.

Due to memory restrictions, setting \strong{sequence.out} to FALSE is recommended for
higher values of k. Sequences corresponding to permutation indexes can easily
be retrieved using function \code{\link[arrangements]{permutations}} from
package \code{\link[arrangements]{arrangements-package}}.

Dinucleotide prefixes and suffixes are represented with numbers 1-16 (corresponding
to their position when ordered alphabetically) to facilitate later matching and
reduce memory requirements.
}
\examples{
LookupTable("con", 1, sequence.out=TRUE)
LookupTable("dnase", 3, sequence.out=TRUE)
LookupTable("con", 5, sequence.out=FALSE)

# retrieving the 1st, 1500th, 10000th and 15000th sequence from the last call:
v <- c(1, 1500, 10000, 15000)
b <- c("A","C","G","T")
apply(arrangements::permutations(b, 5+2, replace=TRUE, index=v), 1, paste, collapse="")
}
